cambios nose

+++ funciones_archivo/manejoCarpetas.py
@@ -1,30 +1,33 @@
 import os, shutil
 
-# Funcion para crear una copia de la carpeta maven en la carpeta del alumno.
-def agregarCarpetaEjercicioEstudiante(rutaArchivador, ejercicio_id, ejercicio_nombre, ejercicio_path):
-    # Si la carpeta ya existe, se elimina y se crea nuevamente
-    rutaEjercicioEstudiante = os.path.join(rutaArchivador, f"Ejercicio_{ejercicio_id}_{ejercicio_nombre}")
-    # Si la carpeta ya existe, seelimina el contenido y se crea nuevamente
+def agregarCarpetaEjercicioEstudiante(rutaArchivador, ejercicio_id, ejercicio_path):
+
+    # Validar parámetros
+    if not os.path.exists(rutaArchivador) or not os.path.exists(ejercicio_path):
+        raise ValueError("Ruta de archivador o ejercicio no válida.")
+
+    # Crear la ruta de la carpeta del ejercicio para el estudiante
+    rutaEjercicioEstudiante = os.path.join(rutaArchivador, f"Ejercicio_{ejercicio_id}")
+
+    # Verificar si la carpeta existe y eliminarla
     if os.path.exists(rutaEjercicioEstudiante):
         shutil.rmtree(rutaEjercicioEstudiante)
-        os.makedirs(rutaEjercicioEstudiante)
-    # Si la carpeta no existe, se crea
-    else:
-        os.makedirs(rutaEjercicioEstudiante)
-    # Se copia la carpeta del ejercicio
-    try:
-        shutil.copytree(ejercicio_path, rutaEjercicioEstudiante)
-        return rutaEjercicioEstudiante
-    except Exception as e:
-        return "Hubo un error en agregar la carpeta del ejercicio"
+    # Copiar la carpeta del ejercicio al estudiante
+    shutil.copytree(ejercicio_path, rutaEjercicioEstudiante)
+
+    return rutaEjercicioEstudiante
 
 def agregarCarpetaSerieEstudiante(matricula, serie_id, serie_nombre):
-    rutaSerieEstudiante= os.path.join("ejerciciosEstudiantes/", f"{matricula}/{serie_id}_{serie_nombre}")
-    if os.path.exists(rutaSerieEstudiante):
-        return rutaSerieEstudiante
-    else:
-        os.makedirs(rutaSerieEstudiante)
-        return rutaSerieEstudiante
+    try:
+        rutaArchivador = crearArchivadorEstudiante(matricula)
+        rutaSerieEstudiante = os.path.join(rutaArchivador, f"{serie_id}_{serie_nombre}")
+        if os.path.exists(rutaSerieEstudiante):
+            return rutaSerieEstudiante
+        else:
+            os.makedirs(rutaSerieEstudiante)
+            return rutaSerieEstudiante
+    except Exception as e:
+        return f"Hubo un error al agregar la carpeta de la serie: {str(e)}"
 
 def crearArchivadorEstudiante(matricula):
 # Funcion crea la carpeta del estudiante con la matricula para guardar sus archivos
@@ -33,11 +36,11 @@ def crearArchivadorEstudiante(matricula):
     if not os.path.exists(rutaEstudiante):
         os.makedirs(rutaEstudiante)
     else:
-        return "Estudiante ya tiene su archivador"
+        return rutaEstudiante
     return rutaEstudiante
 
 # Funcion para crear la carpeta del ejercicio en la carpeta de su respectiva serie.
-def crear_carpeta_ejercicio(id_ejercicio, id_serie, serie_nombre):
+def crearCarpetaEjercicio(id_ejercicio, id_serie, serie_nombre):
     rutaBase = "ejerciciosPropuestos/"
     rutaSerie = os.path.join(rutaBase, f"{id_serie}_{serie_nombre}")
     nombreCarpetaEjercicio = os.path.join(rutaSerie, "Ejercicio_" + str(id_ejercicio))
@@ -56,7 +59,7 @@ def crear_carpeta_ejercicio(id_ejercicio, id_serie, serie_nombre):
 
 
 # Funcion para crear la carpeta de la serie en la carpeta de ejercicios propuestos.
-def crear_carpeta_serie(id_serie, serie_nombre):
+def crearCarpetaSerie(id_serie, serie_nombre):
     # Crea una carpeta para una serie específica
     rutaBase = "ejerciciosPropuestos/"
     rutaEnunciados= "enunciadosEjercicios/"
